pyproject.toml
[project]
name = "mcp-office"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "mcp[cli,fastmcp]>=1.15.0",
    "openpyxl>=3.1.5",
    "python-docx>=1.2.0",
    "python-pptx>=1.0.2",
]
